requirements.txt
## Core runtime
fastapi
uvicorn[standard]
pydantic

## Database
sqlalchemy
python-dotenv

## Messaging
kafka-python

## Scheduling & Task Management
APScheduler

## File handling & CSV
pandas

## AI / ML (core)
torch
transformers
numpy
scipy

## Optional / Hardware-accelerated (install only if you need them)
# optimum-intel
# openvino
# nncf
# accelerate

# Notes:
# - Pin versions as needed for your environment; keep the list minimal for local dev.
# - Optional Intel/OpenVINO packages are commented out — enable them only on supported hardware.
